Enabled 64-bit to 32-bit conversions warning and fixed them

+++ CDLCDyldInfo.h
@@ -37,6 +37,6 @@
 
 // Exported symbols
 - (void)logExportedSymbols;
-- (void)printSymbols:(const uint8_t *)start end:(const uint8_t *)end prefix:(NSString *)prefix offset:(uint32_t)offset;
+- (void)printSymbols:(const uint8_t *)start end:(const uint8_t *)end prefix:(NSString *)prefix offset:(uint64_t)offset;
 
 @end
